+++ packages/bluesky_cli/lib/src/command/commands.dart
@@ -8,6 +8,7 @@ import 'actor/profile.dart';
 import 'actor/profiles.dart';
 import 'actor/search_actors.dart';
 import 'actor/suggestions.dart';
+import 'actor/update_preferences.dart';
 import 'bsky_command.dart';
 import 'common/delete.dart';
 import 'feed/feed.dart';
@@ -42,6 +43,7 @@ final actorCommands = <BskyCommand>[
   SuggestionsCommand(),
   ActorsTypeaheadCommand(),
   PreferencesCommand(),
+  PutPreferences(),
 ];
 
 final feedCommands = <BskyCommand>[