+++ packages/bluesky/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Release Note
 
-## v0.4.7
+## v0.5.0
 
 - Supported `app.bsky.actor.getPreferences` as `findPreferences` in `ActorsService`. ([#387](https://github.com/myConsciousness/atproto.dart/issues/387))
+- Fixed field type to `AtUri` from `String` in `ActorViewer` and `PostViewer`.
+- Supported `app.bsky.actor.putPreferences` as `updatePreferences` in `ActorsService`. ([#388](https://github.com/myConsciousness/atproto.dart/issues/388))
 
 ## v0.4.6