+++ packages/bluesky_cli/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.3
 
 - Supported `preferences` command. ([#387](https://github.com/myConsciousness/atproto.dart/issues/387))
+- Supported `put-preferences` command. ([#388](https://github.com/myConsciousness/atproto.dart/issues/388))
 
 ## v0.1.2